test(protocol): convert smartest agent tests to assertLLM pattern (task 4)
- Migrate opportunity.evaluator stress test (25 unrelated candidates) from runScenario/expectSmartest to assertLLM
- Merge stress test into tests/agents/opportunity.evaluator.spec.ts
- Migrate chat.prompt.dynamic.spec.ts to tests/agents/ replacing runScenario with direct graph.invoke + assertLLM
- Delete src/agents/tests/opportunity.evaluator.smartest.spec.ts
- Fix tests/tsconfig.json: add bun types via typeRoots + types override
